requirements.txt
aiogram>=3.0.0
aiohttp
qrcode
Pillow
python-dotenv
psutil
validators
opencv-python-headless
pyzbar
numpy
cryptography
